+++ R/Emvnorm.R
@@ -1,5 +1,12 @@
-mvnorm.etest <- function(x, R) {
+mvnorm.test <- function(x, R) {
   # parametric bootstrap E-test for multivariate normality
+  if (missing(R)) {
+    method = "Energy test of multivariate normality: (Specify R > 0 for MC test)"
+    R <- 0
+  } else {
+    method = "Energy test of multivariate normality: estimated parameters"
+  }
+  
   if (is.vector(x) || NCOL(x)==1) {
     n <- length(x)
     d <- 1
@@ -20,11 +27,15 @@ mvnorm.etest <- function(x, R) {
 
   names(bootobj$t0) <- "E-statistic"
   e <- list(statistic = bootobj$t0, p.value = p,
-            method = "Energy test of multivariate normality: estimated parameters",
+            method = method,
             data.name = paste("x, sample size ", n, ", dimension ", d, ", replicates ",
                               R, sep = ""))
   class(e) <- "htest"
   e
+}
+
+mvnorm.etest <- function(x, R) {
+  return(mvnorm.test(x, R))
 }
 
 mvnorm.e <- function(x) {
@@ -64,10 +75,19 @@ normal.test <- function(x, method=c("mc", "limit"), R) {
   ## implements the test for for d=1 
   ## Case 4: composite hypothesis
   method <- match.arg(method)
+  estimate <- c(mean(x), sd(x))
+  names(estimate) <- c("mean", "sd")
+  
   if (method == "mc") {
     ## Monte Carlo approach
     if (missing(R)) R <- 0
-    return(mvnorm.etest(x, R=R))
+    e <- energy::mvnorm.etest(x, R=R)
+    e$method <- "Energy test of normality"
+    e$method <- ifelse(R > 0,
+      paste0(e$method,": estimated parameters"),
+        paste0(e$method, "  (Specify R > 0 for MC test)"))
+    e$estimate <- estimate
+    return(e)
   }
   
   ## implement test using asymptotic distribution for p-value